Changes AvailabilityTimeSlots to AvailabilityTimeSlotGenerator.
